% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OR_95CI.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Calculate odds ratios and 95% confidence intervals}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{coef}{Coefficient estimates.}

\item{se}{Standard errors.}

\item{siglevel}{Significance level (e.g., 0.05 for a 95% confidence interval).}

\item{roundto}{Number of decimal places to round the results to.}
}
\value{
A character vector with odds ratios and their 95% confidence intervals.
}
\description{
Calculate odds ratios and 95% confidence intervals
}
\examples{
model1 <- glm(y ~ x1 + x2, family = binomial("logit"), data = toydata)
model1coef <- summary(model1)$coef
OR_95CI(model1coef[,1], model1coef[,2], 0.05, 2)
}
\author{
Amrit Tiwana
}
