% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Toy data
Example data for odds ratio package}
\format{
A data frame with 100 rows and 3 variables:
\describe{
  \item{y}{Binary outcome}
  \item{x1}{Continuous predictor}
  \item{x2}{Binary predictor}
}
}
\source{
adapted from ...
}
\usage{
toydata
}
\description{
Toy data
Example data for odds ratio package
}
\references{
list any references
}
\keyword{datasets}
